+++ WHATSNEW.md
@@ -1,6 +1,7 @@
  + added tumblr.com meta plugin
  + added tumblr.com photo plugin
  + added tumblr.com video plugin
+ + added tumblr.com text plugin
 
 2013.08.06, Version 0.4.2