+++ WHATSNEW.md
@@ -4,6 +4,7 @@
  + added tumblr.com text plugin
  * fixed animoto.com plugin
  * updated guardian.com plguin domain
+ * updated businessinsider.com plguin
 
 2013.08.08, Version 0.4.3