+++ WHATSNEW.md
@@ -6,6 +6,7 @@
  * updated guardian.com plguin domain
  * updated businessinsider.com plguin
  * fixed slideshare.net plugin
+ + added huffingtonpost.com plugin
 
 2013.08.08, Version 0.4.3